+++ shared-data/js/fixtures.ts
@@ -73,7 +73,12 @@ import {
   WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE,
 } from './constants'
 import { getDeckDefFromRobotType } from './helpers'
-import { getAAWithFakesFromCutoutFixtureId } from './helpers/deckConfiguration/getAddressableAreaFrom'
+import {
+  getAAForModuleFixture,
+  getAAsToFixtureIdFromDeckDefWithFakes,
+  getAAWithFakesFromCutoutFixtureId,
+  getMainAAForAFixture,
+} from './helpers/deckConfiguration/getAddressableAreaFrom'
 import {
   getCutoutFixtureIdsForModuleModel,
   getCutoutFixturesForModuleModel,
@@ -127,6 +132,7 @@ export function getAADisplayName(
   ).displayName
 }
 
+// TODO(tz, 12-22-25): Remove this and use deckDefinition instead
 // mapping of OT-2 deck slots to cutouts
 export const OT2_CUTOUT_BY_SLOT_ID: { [slotId: string]: OT2CutoutId } = {
   1: 'cutout1',
@@ -143,6 +149,7 @@ export const OT2_CUTOUT_BY_SLOT_ID: { [slotId: string]: OT2CutoutId } = {
   fixedTrash: 'cutout12',
 }
 
+// TODO(tz, 12-22-25): Remove this and use deckDefinition instead
 // mapping of Flex deck slots to cutouts
 export const FLEX_CUTOUT_BY_SLOT_ID: { [slotId: string]: CutoutId } = {
   A1: 'cutoutA1',
@@ -685,6 +692,7 @@ export function getFixtureDisplayName(
   }
 }
 
+// TODO(tz, 12-22-25): Remove this and use deckDefinition instead
 // TODO: Move to helpers/deckDeclarationHelpers.ts
 export const STANDARD_OT2_SLOTS: AddressableAreaName[] = [
   ADDRESSABLE_AREA_1,
@@ -700,6 +708,7 @@ export const STANDARD_OT2_SLOTS: AddressableAreaName[] = [
   ADDRESSABLE_AREA_11,
 ]
 
+// TODO(tz, 12-22-25): Remove this and use deckDefinition instead
 // TODO: Move to helpers
 export const STANDARD_FLEX_SLOTS: AddressableAreaName[] = [
   A1_ADDRESSABLE_AREA,
@@ -724,95 +733,6 @@ export const isAddressableAreaStandardSlot = (
     ? STANDARD_FLEX_SLOTS
     : STANDARD_OT2_SLOTS
   ).includes(addressableAreaName)
-
-/**
- * Given a cutout id get a key value pair of all possibilities for fixture id and related AA
- * @param addedCutoutConfigs: fixtures list selected to add to deck
- * @param cutoutId: cutout if we are adding a fixture to.
- * @returns key value pair of of all possibilities for fixture id and related AA
- */
-export const getAAsToFixtureIdFromDeckDefWithFakes = (
-  cutoutId: CutoutId,
-  deckDef: DeckDefinition
-): Record<CutoutFixtureIdsWithFakes, AddressableAreaNamesWithFakes[]> => {
-  const deckDefWithFakes = getDeckDefWithFakes(deckDef)
-  // replace staging area aaId to fake ones
-  const availableCutoutFixtuers = deckDefWithFakes.cutoutFixtures.filter(cf =>
-    cf.mayMountTo.includes(cutoutId)
-  )
-  const aaForCutoutFixrure = availableCutoutFixtuers.reduce<
-    Partial<Record<CutoutFixtureIdsWithFakes, AddressableAreaNamesWithFakes[]>>
-  >((acc, { id, providesAddressableAreas }) => {
-    acc[id] = providesAddressableAreas[cutoutId]
-    return acc
-  }, {})
-  return aaForCutoutFixrure as Record<
-    CutoutFixtureIdsWithFakes,
-    AddressableAreaNamesWithFakes[]
-  >
-}
-
-export const getAAForModuleFixture = (
-  cutoutId: CutoutId,
-  fixtureId: CutoutFixtureIdsWithFakes,
-  moduleModel: ModuleModel
-): AddressableAreaNamesWithFakes => {
-  const deckDef = getDeckDefFromRobotType(FLEX_ROBOT_TYPE)
-  const aaList = getAAWithFakesFromCutoutFixtureId(cutoutId, fixtureId, deckDef)
-  const aa = aaList?.find(
-    aa =>
-      getAAByAAId(aa, deckDef).areaType ===
-      FLEX_MODULE_AA_TYPE_BY_MODEL[moduleModel]
-  )
-  if (!aa) {
-    console.error(
-      `Was not able to find a aa match for module ${moduleModel} in cutout ${cutoutId}`
-    )
-    return A1_ADDRESSABLE_AREA
-  }
-  return aa
-}
-
-/**
- * get relevent aa name that match with cutoutId and fixtureId.
- *
- * @param cutoutId - The cutoutId we are looking for.
- * @param fixtureId - The fixtureId we are looking for.
- * @returns The aa name or null if not match found.
- */
-export const getMainAAForAFixture = (
-  cutoutId: CutoutId,
-  fixtureId: CutoutFixtureId,
-  addressableAreaId: AddressableAreaNamesWithFakes,
-  existingCutoutFixtureId?: CutoutFixtureIdsWithFakes
-): AddressableAreaNamesWithFakes | null => {
-  const addressableAreasByFixtureId = getAAsToFixtureIdFromDeckDefWithFakes(
-    cutoutId,
-    getDeckDefFromRobotType('OT-3 Standard')
-  )
-  const aaListForFixtureId = addressableAreasByFixtureId[fixtureId] ?? []
-  if (LEFT_AND_CENTER_CUTOUTS.includes(cutoutId)) {
-    return aaListForFixtureId[0]
-  } else if (WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE === fixtureId) {
-    return DEFAULT_AA_FOR_WASTE_CHUTE
-  } else if (fixtureId === TRASH_BIN_ADAPTER_FIXTURE) {
-    if (
-      existingCutoutFixtureId &&
-      COMBO_FIXTURES.includes(existingCutoutFixtureId)
-    ) {
-      return null
-    } else {
-      return aaListForFixtureId[0]
-    }
-  } else {
-    const aa = aaListForFixtureId.find((aa: AddressableAreaNamesWithFakes) => {
-      const vsId = getVisualSlotIdFromAAId(aa)
-      const singleSlotId = getAAWithFakesFromVSId(vsId)
-      return singleSlotId === addressableAreaId
-    })
-    return aa! // we can cast this bc there should me a match for every fixtureId
-  }
-}
 
 export const getMainUsbModuleFixtureIdForComboFixture = (
   compatibleCutoutFixtureIds: CutoutFixtureId[]
